% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get URL Data}
\usage{
getData(url)
}
\arguments{
\item{url:}{web URL}
}
\value{
list (for json based object) or text
}
\description{
getData function accepts URL parameter, based on content Type it returns list or text formated
for content.type=="application/json"  = list data returned
getData function also cache the responce data to optimize the performance
only first time URL actualy requested, but after subsequent call URL reponce pulled from cache
}
\examples{
 q<-getData("http://echo.jsontest.com/fieldkey/fieldvalue/purpose/test")
 q<-getData("http://www.google.com")
}
